# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact: {}
  description: Auth Service for B2B Suite – admin, employee, attendance, etc.
  title: B2B Suite API
  version: "1.0"
paths:
...
  /api/v1/admin/admins/{adminID}/avatar:
    delete:
      parameters:
      - description: Admin UUID
        in: path
        name: adminID
        required: true
        type: string
      responses:
        "200":
          description: Avatar removed
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid admin ID
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
      summary: Deactivate admin avatar
      tags:
      - admin-avatar
    get:
      parameters:
      - description: Admin UUID
        in: path
        name: adminID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Avatar details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid admin ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Avatar not found
          schema:
            additionalProperties: true
            type: object
      summary: Get admin avatar
      tags:
      - admin-avatar
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin UUID
        in: path
        name: adminID
        required: true
        type: string
      - description: Avatar data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Avatar set
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
      summary: Set admin avatar
      tags:
      - admin-avatar
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"